requirements.txt
absl-py==1.1.0
astunparse==1.6.3
boto3==1.24.24
botocore==1.27.24
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dill==0.3.5.1
etils==0.6.0
Flask==2.1.2
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
gast==0.5.3
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.3
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
mysqlclient==1.4.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
promise==2.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
six==1.15.0
SQLAlchemy==1.4.39
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metadata==1.9.0
tensorflow-metal==0.5.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
